NXP-28023: fix testComputationRecoveryPolicy should not be run on multiple log implementation
